+++ src/main/java/com/sparta/harmony/common/config/OpenApiConfig.java
@@ -1,4 +1,4 @@
-package com.sparta.harmony.order.config;
+package com.sparta.harmony.common.config;
 
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Info;
