+++ src/main/java/com/sparta/harmony/common/handler/success/SuccessResponseHandler.java
@@ -1,6 +1,7 @@
-package com.sparta.harmony.order.handler.success;
+package com.sparta.harmony.common.handler.success;
 
-import com.sparta.harmony.order.dto.*;
+import com.sparta.harmony.common.dto.ApiResponseDto;
+import com.sparta.harmony.common.dto.ApiResponsePageDto;
 import org.springframework.data.domain.Page;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
